# Allow relying on Jackson-databind in Dev UI JsonRpc services regardless of runtime dependencies