… extensions/vertx-http/runtime/src/main/java/io/quarkus/devui/runtime/jsonrpc/JsonRpcWriter.java
… extensions/vertx-http/runtime/src/main/java/io/quarkus/devui/runtime/jsonrpc/json/JsonMapper.java
… extensions/vertx-http/runtime/src/main/java/io/quarkus/devui/runtime/jsonrpc/json/JsonTypeAdapter.java
… extensions/vertx/runtime/src/main/java/io/quarkus/vertx/runtime/jackson/BufferDeserializer.java
… extensions/vertx/runtime/src/main/java/io/quarkus/vertx/runtime/jackson/ByteArrayDeserializer.java
… extensions/vertx/runtime/src/main/java/io/quarkus/vertx/runtime/jackson/ByteArraySerializer.java
… extensions/vertx/runtime/src/main/java/io/quarkus/vertx/runtime/jackson/InstantDeserializer.java
… extensions/vertx/runtime/src/main/java/io/quarkus/vertx/runtime/jackson/InstantSerializer.java
… extensions/vertx/runtime/src/main/java/io/quarkus/vertx/runtime/jackson/JsonUtil.java