This will allow any JsonRPC service involved in Dev UI (provided through `JsonRPCProvidersBuildItem`) to take advantage of Jackson-Databind, thereby removing the need to manually serialize/deserialize using `JsonObject`/`JsonArray`/etc.: you can just model your data using POJOs and have it converted to JSON automatically, like we do in `HibernateOrmDevJsonRpcService#getInfo`. This will work even if the application's runtime classpath doesn't include Jackson-databind.

Of course, you can *still* manually serialize/deserialize using `JsonObject`/`JsonArray`/etc if that's your cup of tea; that still works.

Limitations:

* POJOs will be serialized as-is, deriving JSON property names from the name of public fields/getters, without any way to configure serialization/deserialization any further:
  * The Jackson-databind `ObjectMapper` is not configurable (though we could potentially allow that through build items).
  * The Jackson-databind `ObjectMapper` that performs serialization lives in the deployment classloader, so even if you do have jackson-databind in your runtime classpath, the object mapper will probably not recognize Jackson-databind annotations on your POJOs (I imagine they would be seen as instances of different annotations).
* JsonRPC services must still be implemented in the runtime module. I thought that would be simpler given we often get information from the runtime and transferring data from the runtime classloader to the deployment classloader is a real pain.

I mainly tested the DevUI for the Hibernate extensions and for config, which seem to work fine, even with Jackson-databind absent from the runtime classpath. So this PR at least fixes the bug where the Hibernate DevUIs were not working at all without jackson-databind in your runtime classpath.

In the absence of automated tests for the new Dev UI, I'm open to suggestions as to what to test and how.