@yrodiere 2023-04-04
> Do you have any measurements of how this affects startup time in dev mode?

No actual benchmarks, if that's what you're asking. That didn't even cross my mind, to be honest: just like I would not think reasonable to code in assembly for better performance, I don't think it's reasonable to serialize stuff manually for better performance.

But FWIW:

* We were already using jackson before this change; I didn't have to add a dependency.
* The main overhead on startup would thus be... the creation of the Jackson `ObjectMapper`.
* On my machine, with a simple application with Hibernate ORM + Hibernate Search + Postgres + Elasticsearch, devservices disabled, schema management disabled:
  * Main branch: ~3.2s inital start, ~0.5s restart
  * This PR: ~3.2s inital start, ~0.5s restart