@geoand 2023-04-04
My concern is the additional creation of the ObjectMapper.
If the difference is tiny, then just ignore my comment 🙂